backend/requirements.txt
fastapi[standard]
uvicorn[standard]
python-dotenv
pydantic-settings
supabase
openai-whisper
groq
python-multipart
